perf: scalar replacement of non-escaping objects + Static Hermes benchmarks
Escape analysis identifies `let p = new Point(x, y)` where `p` never
escapes (only used in PropertyGet/PropertySet). Fields are decomposed
into stack allocas that LLVM promotes to registers — zero heap allocation.

object_create: 10ms → 4ms (2.5x), binary_trees: 9ms → 3ms (3x).
Perry now beats Node.js on all 15 benchmarks (was 13/15).
Peak RSS for binary_trees dropped from 97MB to 5MB.

Also adds Static Hermes (Meta) to the benchmark suite runner with
automatic type-stripping for the TS→JS conversion. Updated README
with full 4-way comparison tables (Perry vs Node vs Bun vs Hermes)
and refreshed polyglot numbers.

Removes accidentally committed benchmark binaries (~42MB).

+++ crates/perry-runtime/src/array.rs
@@ -11,7 +11,7 @@ use crate::arena::arena_alloc_gc;
 /// Strip NaN-boxing tags from an array pointer and guard against invalid values.
 #[inline(always)]
 fn clean_arr_ptr(arr: *const ArrayHeader) -> *const ArrayHeader {
-    let bits = arr as usize;
+    let bits = arr as u64;
     let top16 = bits >> 48;
     if top16 >= 0x7FF8 {
         if top16 == 0x7FFC || (bits & 0x0000_FFFF_FFFF_FFFF) == 0 {
